WIP: Switch to single script
